+++ frontend/src/API/UserService/editProfile.ts
@@ -5,6 +5,6 @@ export const editProfile = async (data: {}, profileID: string | undefined) => {
         const response = axiosInstance.patch(`/profiles/${profileID}`, JSON.stringify(data));
         return response;
     } catch (error) {
-        throw new Error("Failed");
+        throw new Error("Editing profile failed!");
     }
 };
