+++ frontend/src/components/Modal/Modal.module.scss
@@ -11,15 +11,13 @@
     right: 0;
     width: 100vw;
     height: 100vh;
-    background-color: rgba($color: #262626, $alpha: 0.5);
-    z-index: 9999;
+    background-color: rgba($color: #262626, $alpha: 0.7);
+    z-index: 10;
 }
 
 .exitIcon {
-    width: "14px";
     border: 1px solid $c-neutrals-6;
     border-radius: 50%;
-    padding: 5px;
 }
 .exitIcon:hover {
     cursor: pointer;
@@ -38,8 +36,7 @@
 
 .modal_header {
     width: 100%;
-    border-bottom: 1px solid var(--c-neutral-20);
-    padding-bottom: var(--spacer-16);
+    padding-bottom: $spacer-16;
 }
 
 .modal_text {
@@ -52,7 +49,7 @@
 }
 
 .modal_description {
-    margin-top: 8px;
+    margin-top: $spacer-8;
     width: 100%;
 }
