+++ frontend/src/components/navbarDropDown/NavbarDropDown.tsx
@@ -11,7 +11,8 @@ import ArrowDownSign from "../../assets/ArrowDownSign.svg";
 import profileIcon from "../../assets/profileIcon.svg";
 import plusIcon from "../../assets/plusIcon.svg";
 import { Link } from "react-router-dom";
-import { changeAddProfileVisibility } from "store/addProfileModal";
+import Modal from "components/Modal/Modal";
+import { ProfileModal } from "components/Modal/ProfileModal/ProfileModal";
 
 export function NavbarDropDown() {
     const dispatch = useAppDispatch();
@@ -20,6 +21,7 @@ export function NavbarDropDown() {
     const activeAccount = useAppSelector((state) => state.activeProfile);
     const [activeDropdownButton, setActiveDropdownButton] = useState(false);
     const [activeProfileSwitchButton, setActiveProfileSwitchButton] = useState(false);
+    const [isModalVisible, setIsModalVisible] = useState(false);
 
     return (
         <div className={styles.dropdownMenu}>
@@ -91,7 +93,7 @@ export function NavbarDropDown() {
                     <button
                         type="button"
                         className={styles.profilesListItem}
-                        onClick={() => dispatch(changeAddProfileVisibility(true))}
+                        onClick={() => setIsModalVisible(true)}
                     >
                         <img src={plusIcon} alt="plus" className={styles.dropdownListIcon} />
                         Create new profile
@@ -109,6 +111,17 @@ export function NavbarDropDown() {
                     Disconnect
                 </button>
             </div>
+            {isModalVisible && (
+                <Modal
+                    visible={isModalVisible}
+                    onClose={() => {
+                        setIsModalVisible(false);
+                    }}
+                    title="Create new profile"
+                >
+                    <ProfileModal isNew={true} userID={user.userID} />
+                </Modal>
+            )}
         </div>
     );
 }
