+++ frontend/src/store/store.ts
@@ -4,8 +4,6 @@ import userReducer from "./user";
 import profilesReducer from "./profile";
 import auctionsReducer from "./auctions";
 import activeProfileReducer from "./activeProfile";
-import addProfileModalVisibilityReducer from "./addProfileModal";
-import editProfileModalVisibilityReducer from "./editProfileModal";
 
 export const store = configureStore({
     reducer: {
@@ -13,8 +11,6 @@ export const store = configureStore({
         profiles: profilesReducer,
         auctions: auctionsReducer,
         activeProfile: activeProfileReducer,
-        addProfileModalVisibility: addProfileModalVisibilityReducer,
-        editProfileModalVisibility: editProfileModalVisibilityReducer,
     },
 });
